Implement REST API (#30)
* new logic in renko display - not satisfied

* Added the required Rest API to the application.

* Removed backup code

* Renamed the DTO Object

+++ backend/src/main/java/com/prem/ta/entities/Ticker.java
@@ -12,7 +12,7 @@ public class Ticker {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer tickerId;
+    private Long tickerId;
 
     @Column(name = "start_date", nullable = false)
     private LocalDate startDate;
